moved class to it's own file
